05-05

+++ Dsms/admin/Deleteproduct.aspx.cs
@@ -12,6 +12,7 @@ namespace Dsms.admin
     public partial class Deleteproduct : System.Web.UI.Page
     {
         SqlConnection con = new SqlConnection(ConfigurationManager.ConnectionStrings["dbconnection"].ConnectionString);
+        
         protected void Page_Load(object sender, EventArgs e)
         {
             Response.Cache.SetExpires(DateTime.UtcNow.AddDays(-1));
@@ -25,10 +26,20 @@ namespace Dsms.admin
             {
                 con.Open();
                 string id = Request.QueryString["pid"];
-                string delete = "delete from tblProduct where pid='" + id + "'";
-                SqlCommand com = new SqlCommand(delete, con);
-                com.ExecuteNonQuery();
-                Response.Redirect("product.aspx");
+                string query = "select * from tblProduct where pid='" + id + "'";
+                SqlCommand cmd = new SqlCommand(query,con);
+                SqlDataReader dr = cmd.ExecuteReader();
+                if (dr.Read())
+                {
+                    string stock = "delete from tblStock where pname='" + dr.GetString(1) + "'";
+                    SqlCommand st = new SqlCommand(stock, con);
+                    st.ExecuteNonQuery();
+                    string delete = "delete from tblProduct where pid='" + id + "'";
+                    SqlCommand com = new SqlCommand(delete, con);
+                    com.ExecuteNonQuery();
+                    Response.Redirect("product.aspx");
+                }
+                
             }
             else
             {
